fix(export): prepend UTF-8 BOM to CSV so Excel renders CJK
Excel on Chinese Windows falls back to the GBK code page for BOM-less UTF-8 CSV and garbles CJK. buildCsvContent now prepends U+FEFF. Extracted to csvExport.ts (pure, testable) and added vitest + BOM/escaping tests.
